Creating Multi-Threaded TDCM Stan Code

Automating the creation of multi-threaded Stan code for a TDCM.

create_threaded_stan_tdcm(q_matrix, profs)

Arguments

q_matrix

A tibble containing the assessment Q-matrix.

profs

A tibble of the possible attribute mastery profiles.

Value

`stan_code` A list containing the text for the Stan code blocks.

Examples

qmatrix = tibble::tibble(att_1 = c(1, 0, 1, 0, 1, 1), att_2 = c(0, 1, 0, 1, 1, 1))
possible_profiles = tibble::tibble(att_1 = c(0, 1, 0, 1), att_2 = c(0, 1, 0, 1))
create_threaded_stan_tdcm(q_matrix = qmatrix, profs = possible_profiles)
#> $functions
#> functions {
#>   real minmax (real x) {
#>     if (x < .01) {
#>       return 0.01;
#>     }
#> 
#>     if (x > 0.99) {
#>       return 0.99;
#>     }
#> 
#>     return x;
#>   }
#> 
#>   vector sum_probs(vector beta, vector theta, real[] xr, int[] xi) {
#>     int Z = num_elements(xi);
#>     int ys = xi[Z - 1];
#>     int iis = xi[Z];
#> 
#>     int y1[iis] = xi[1:iis];
#>     int ii1[iis] = xi[(iis + 1):(2 * iis)];
#>     int jj1[iis] = xi[((2 * iis) + 1):(3 * iis)];
#>     int s1[ys] = xi[((3 * iis) + 1):((3 * iis) + ys)];
#>     int l1[ys] = xi[((3 * iis) + ys + 1):((3 * iis) + (2 * ys))];
#> 
#>     int y2[iis] = xi[((3 * iis) + (2 * ys) + 1):((4 * iis) + (2 * ys))];
#>     int ii2[iis] = xi[((4 * iis) + (2 * ys) + 1):((5 * iis) + (2 * ys))];
#>     int jj2[iis] = xi[((5 * iis) + (2 * ys) + 1):((6 * iis) + (2 * ys))];
#>     int s2[ys] = xi[((6 * iis) + (2 * ys) + 1):((6 * iis) + (3 * ys))];
#>     int l2[ys] = xi[((6 * iis) + (3 * ys) + 1):((6 * iis) + (4 * ys))];
#> 
#>     int I = xi[Z - 7];
#>     int N = xi[Z - 6];
#>     int C = xi[Z - 5];
#>     int A = xi[Z - 4];
#>     int J = xi[Z - 3];
#>     int M = xi[Z - 2];
#> 
#>     vector[C] Vc = beta[1:C];
#>     vector[I * C] pic = beta[(C + 1):(C + (I * C))];
#>     vector[C * C] tauc = beta[(C + (I * C) + 1):(C + (I * C) + (C * C))];
#> 
#>     matrix[C, C] ps;
#>     matrix[C, C] tau_c;
#>     real person = 0;
#> 
#>     matrix[I, C] pi_c;
#>     for(c in 1:C) {
#>       for(i in 1:I) {
#>         int ic = i + ((c-1) * I);
#>         pi_c[i, c] = pic[ic];
#>       }
#>     }
#> 
#>     for(c1 in 1:C) {
#>       for(c2 in 1:C) {
#>         int cc = c1 + ((c2 - 1) * C);
#>         tau_c[c1, c2] = tauc[cc];
#>       }
#>     }
#> 
#>     // Likelihood
#>     for (j in 1:J) {
#>       vector[C] tmp;
#>       for (c1 in 1:C) {
#>         for (c2 in 1:C) {
#>           real log_items[l1[j]];
#>           for (m in 1:l1[j]) {
#>             int i = ii1[s1[j] + m - 1];
#>             log_items[m] = y1[s1[j] + m - 1] * log(pi_c[i,c1]) + (1 - y1[s1[j] + m - 1]) * log(1 - pi_c[i,c1]) + y2[s1[j] + m - 1] * log(pi_c[i,c2]) + (1 - y2[s1[j] + m - 1]) * log(1 - pi_c[i,c2]);
#>           }
#>           ps[c1, c2] = log(Vc[c1]) + log(tau_c[c1, c2]) + sum(log_items);
#>         }
#>         tmp[c1] = log_sum_exp(ps[c1,]);
#>       }
#>       person += log_sum_exp(tmp);
#>     }
#> 
#>     return [person]';
#>   }
#> 
#>   vector person_loglik(vector beta, vector theta, real[] xr, int[] xi) {
#>     int Z = num_elements(xi);
#>     int ys = xi[Z - 1];
#>     int iis = xi[Z];
#> 
#>     int y1[iis] = xi[1:iis];
#>     int ii1[iis] = xi[(iis + 1):(2 * iis)];
#>     int jj1[iis] = xi[((2 * iis) + 1):(3 * iis)];
#>     int s1[ys] = xi[((3 * iis) + 1):((3 * iis) + ys)];
#>     int l1[ys] = xi[((3 * iis) + ys + 1):((3 * iis) + (2 * ys))];
#> 
#>     int y2[iis] = xi[((3 * iis) + (2 * ys) + 1):((4 * iis) + (2 * ys))];
#>     int ii2[iis] = xi[((4 * iis) + (2 * ys) + 1):((5 * iis) + (2 * ys))];
#>     int jj2[iis] = xi[((5 * iis) + (2 * ys) + 1):((6 * iis) + (2 * ys))];
#>     int s2[ys] = xi[((6 * iis) + (2 * ys) + 1):((6 * iis) + (3 * ys))];
#>     int l2[ys] = xi[((6 * iis) + (3 * ys) + 1):((6 * iis) + (4 * ys))];
#> 
#>     int I = xi[Z - 7];
#>     int N = xi[Z - 6];
#>     int C = xi[Z - 5];
#>     int A = xi[Z - 4];
#>     int J = xi[Z - 3];
#>     int M = xi[Z - 2];
#> 
#>     vector[C] Vc = beta[1:C];
#>     vector[I * C] pic = beta[(C + 1):(C + (I * C))];
#>     vector[C * C] tauc = beta[(C + (I * C) + 1):(C + (I * C) + (C * C))];
#> 
#>     matrix[C, C] ps;
#>     matrix[C, C] tau_c;
#>     vector[J] person;
#> 
#>     matrix[I, C] pi_c;
#>     for(c in 1:C) {
#>       for(i in 1:I) {
#>         int ic = i + ((c-1) * I);
#>         pi_c[i, c] = pic[ic];
#>       }
#>     }
#> 
#>     for(c1 in 1:C) {
#>       for(c2 in 1:C) {
#>         int cc = c1 + ((c2 - 1) * C);
#>         tau_c[c1, c2] = tauc[cc];
#>       }
#>     }
#> 
#>     // Likelihood
#>     for (j in 1:J) {
#>       vector[C] tmp;
#>       for (c1 in 1:C) {
#>         for (c2 in 1:C) {
#>           real log_items[l1[j]];
#>           for (m in 1:l1[j]) {
#>             int i = ii1[s1[j] + m - 1];
#>             log_items[m] = y1[s1[j] + m - 1] * log(pi_c[i,c1]) + (1 - y1[s1[j] + m - 1]) * log(1 - pi_c[i,c1]) + y2[s1[j] + m - 1] * log(pi_c[i,c2]) + (1 - y2[s1[j] + m - 1]) * log(1 - pi_c[i,c2]);
#>           }
#>           ps[c1, c2] = log(Vc[c1]) + log(tau_c[c1, c2]) + sum(log_items);
#>         }
#>         tmp[c1] = log_sum_exp(ps[c1,]);
#>       }
#>       person[j] = log_sum_exp(tmp);
#>     }
#> 
#>     return person;
#>   }
#> 
#>   vector resp_transition(vector beta, vector theta, real[] xr, int[] xi) {
#>     int Z = num_elements(xi);
#>     int ys = xi[Z - 1];
#>     int iis = xi[Z];
#> 
#>     int y1[iis] = xi[1:iis];
#>     int ii1[iis] = xi[(iis + 1):(2 * iis)];
#>     int jj1[iis] = xi[((2 * iis) + 1):(3 * iis)];
#>     int s1[ys] = xi[((3 * iis) + 1):((3 * iis) + ys)];
#>     int l1[ys] = xi[((3 * iis) + ys + 1):((3 * iis) + (2 * ys))];
#> 
#>     int y2[iis] = xi[((3 * iis) + (2 * ys) + 1):((4 * iis) + (2 * ys))];
#>     int ii2[iis] = xi[((4 * iis) + (2 * ys) + 1):((5 * iis) + (2 * ys))];
#>     int jj2[iis] = xi[((5 * iis) + (2 * ys) + 1):((6 * iis) + (2 * ys))];
#>     int s2[ys] = xi[((6 * iis) + (2 * ys) + 1):((6 * iis) + (3 * ys))];
#>     int l2[ys] = xi[((6 * iis) + (3 * ys) + 1):((6 * iis) + (4 * ys))];
#> 
#>     int I = xi[Z - 7];
#>     int N = xi[Z - 6];
#>     int C = xi[Z - 5];
#>     int A = xi[Z - 4];
#>     int J = xi[Z - 3];
#>     int M = xi[Z - 2];
#> 
#>     vector[C] Vc = beta[1:C];
#>     vector[I * C] pic = beta[(C + 1):(C + (I * C))];
#>     vector[C * C] tauc = beta[(C + (I * C) + 1):(C + (I * C) + (C * C))];
#> 
#>     matrix[C, C] ps;
#>     matrix[C, C] tau_c;
#>     matrix[C, C] prob_transition_class[J];
#> 
#>     vector[J * C * C] person;
#> 
#>     matrix[I, C] pi_c;
#>     for(c in 1:C) {
#>       for(i in 1:I) {
#>         int ic = i + ((c-1) * I);
#>         pi_c[i, c] = pic[ic];
#>       }
#>     }
#> 
#>     for(c1 in 1:C) {
#>       for(c2 in 1:C) {
#>         int cc = c1 + ((c2 - 1) * C);
#>         tau_c[c1, c2] = tauc[cc];
#>       }
#>     }
#> 
#>     // latent class probabilities
#>     for (j in 1:J) {
#>       vector[C] tmp;
#>       matrix[C, C] prob_joint;
#>       for (c1 in 1:C) {
#>         for (c2 in 1:C) {
#>           real log_items[l1[j]];
#>           for (m in 1:l1[j]) {
#>             int i = ii1[s1[j] + m - 1];
#>             log_items[m] = y1[s1[j] + m - 1] * log(pi_c[i,c1]) + (1 - y1[s1[j] + m - 1]) * log(1 - pi_c[i,c1]) + y2[s1[j] + m - 1] * log(pi_c[i,c2]) + (1 - y2[s1[j] + m - 1]) * log(1 - pi_c[i,c2]);
#>           }
#>           prob_joint[c1, c2] = log(Vc[c1]) + log(tau_c[c1, c2]) + sum(log_items);
#>         }
#>       }
#>       prob_transition_class[j] = exp(prob_joint) / sum(exp(prob_joint));
#> 
#>       for (c1 in 1:C) {
#>         for (c2 in 1:C) {
#>           person[((c1 - 1) * 8) + c2 + ((j - 1) * C * C)] = prob_transition_class[j,c1,c2];
#>         }
#>       }
#>     }
#>     return person;
#>   }
#> }
#> 
#> $data
#> data {
#>   int<lower=1> I;                    // number of items
#>   int<lower=1> J;                    // number of respondents
#>   int<lower=1> N;                    // number of observations
#>   int<lower=1> C;                    // number of classes
#>   int<lower=1> A;                    // number of attributes
#>   int<lower=1,upper=I> ii[N, 2];     // item for obs n
#>   int<lower=1,upper=J> jj[N, 2];     // respondent for obs n
#>   int<lower=0,upper=1> y[N, 2];      // score for obs n
#>   int<lower=1,upper=N> s[J, 2];      // starting row for j
#>   int<lower=1,upper=I> l[J, 2];      // number of items for j
#>   matrix[C,A] Alpha;                 // attribute pattern for each C
#>   int<lower=1> n_shards;             // the number of shards to split the data into
#> }
#> 
#> $transformed_data
#> transformed data {
#>   int ys = num_elements(s) / 2 / n_shards;
#>   int iis = num_elements(ii) / 2 / n_shards;
#> 
#>   int M = iis;
#> 
#>   int xi[n_shards, (4 * ys) + (6 * iis) + 8];
#> 
#>   // an empty set of per-shard parameters
#>   vector[0] theta[n_shards];
#> 
#>   real xr[n_shards,1];
#>   for(kk in 1:n_shards) {
#>     xr[kk, 1] = 1.0;
#>   }
#> 
#>   // split into shards
#>   for (i in 1:n_shards) {
#>     int ylower;
#>     int yupper;
#>     int iilower;
#>     int iiupper;
#> 
#>     ylower = ((i - 1) * ys) + 1;
#>     yupper = i * ys;
#>     iilower = ((i - 1) * iis) + 1;
#>     iiupper = i * iis;
#> 
#>     xi[i, 1:iis] = y[iilower:iiupper, 1];
#>     xi[i, (iis + 1):(iis + iis)] = ii[iilower:iiupper, 1];
#>     xi[i, ((2 * iis) + 1):((2 * iis) + iis)] = jj[iilower:iiupper, 1];
#>     xi[i, ((3 * iis) + 1):((3 * iis) + ys)] = s[1:ys, 1];
#>     xi[i, ((3 * iis) + ys + 1):((3 * iis) + (2 * ys))] = l[ylower:yupper, 1];
#>     xi[i, ((3 * iis) + (2 * ys) + 1):((4 * iis) + (2 * ys))] = y[iilower:iiupper, 2];
#>     xi[i, ((4 * iis) + (2 * ys) + 1):((5 * iis) + (2 * ys))] = ii[iilower:iiupper, 2];
#>     xi[i, ((5 * iis) + (2 * ys) + 1):((6 * iis) + (2 * ys))] = jj[iilower:iiupper, 2];
#>     xi[i, ((6 * iis) + (2 * ys) + 1):((6 * iis) + (3 * ys))] = s[1:ys, 2];
#>     xi[i, ((6 * iis) + (3 * ys) + 1):((6 * iis) + (4 * ys))] = l[ylower:yupper, 2];
#>     xi[i, ((6 * iis) + (4 * ys) + 1)] = I;
#>     xi[i, ((6 * iis) + (4 * ys) + 2)] = N / n_shards;
#>     xi[i, ((6 * iis) + (4 * ys) + 3)] = C;
#>     xi[i, ((6 * iis) + (4 * ys) + 4)] = A;
#>     xi[i, ((6 * iis) + (4 * ys) + 5)] = J / n_shards;
#>     xi[i, ((6 * iis) + (4 * ys) + 6)] = iis;
#>     xi[i, ((6 * iis) + (4 * ys) + 7)] = ys;
#>     xi[i, ((6 * iis) + (4 * ys) + 8)] = iis;
#>   }
#> }
#> 
#> $parameters
#> parameters {
#>   simplex[C] tau[C];
#>   simplex[C] Vc;
#>   real l1_0;
#>   real l2_0;
#>   real l3_0;
#>   real l4_0;
#>   real l5_0;
#>   real l6_0;
#>   real<lower=0> l1_11;
#>   real<lower=0> l2_12;
#>   real<lower=0> l3_11;
#>   real<lower=0> l4_12;
#>   real<lower=0> l5_11;
#>   real<lower=0> l5_12;
#>   real<lower=0> l6_11;
#>   real<lower=0> l6_12;
#>   real<lower=-1 * fmin(l5_11, l5_12)> l5_212;
#>   real<lower=-1 * fmin(l6_11, l6_12)> l6_212;
#> }
#> 
#> $transformed_parameters
#> transformed parameters {
#>   matrix[I,C] pi;
#> 
#>   pi[1,1] = inv_logit(l1_0);
#>   pi[2,1] = inv_logit(l2_0);
#>   pi[3,1] = inv_logit(l3_0);
#>   pi[4,1] = inv_logit(l4_0);
#>   pi[5,1] = inv_logit(l5_0);
#>   pi[6,1] = inv_logit(l6_0);
#>   pi[1,2] = inv_logit(l1_0+l1_11);
#>   pi[2,2] = inv_logit(l2_0+l2_12);
#>   pi[3,2] = inv_logit(l3_0+l3_11);
#>   pi[4,2] = inv_logit(l4_0+l4_12);
#>   pi[5,2] = inv_logit(l5_0+l5_11+l5_12+l5_212);
#>   pi[6,2] = inv_logit(l6_0+l6_11+l6_12+l6_212);
#>   pi[1,3] = inv_logit(l1_0);
#>   pi[2,3] = inv_logit(l2_0);
#>   pi[3,3] = inv_logit(l3_0);
#>   pi[4,3] = inv_logit(l4_0);
#>   pi[5,3] = inv_logit(l5_0);
#>   pi[6,3] = inv_logit(l6_0);
#>   pi[1,4] = inv_logit(l1_0+l1_11);
#>   pi[2,4] = inv_logit(l2_0+l2_12);
#>   pi[3,4] = inv_logit(l3_0+l3_11);
#>   pi[4,4] = inv_logit(l4_0+l4_12);
#>   pi[5,4] = inv_logit(l5_0+l5_11+l5_12+l5_212);
#>   pi[6,4] = inv_logit(l6_0+l6_11+l6_12+l6_212);
#> 
#>   real pic[I * C];
#>   for(c in 1:C) {
#>     for(i in 1:I) {
#>       int ic = i + ((c - 1) * I);
#>       pic[ic] = pi[i, c];
#>     }
#>   }
#> 
#>   real tauc[C * C];
#>   for(c1 in 1:C) {
#>     for(c2 in 1:C) {
#>       int cc = c2 + ((c1 - 1) * C);
#>       tauc[cc] = tau[c2, c1];
#>     }
#>   }
#> 
#>   // a set of shared parameters
#>   vector[C + (I * C) + (C * C)] beta;
#>   beta[1:C] = Vc[1:C];
#>   beta[(C + 1):(C + (I * C))] = to_vector(pic[1:(I * C)]);
#>   beta[(C + (I * C) + 1):(C + (I * C) + (C * C))] = to_vector(tauc[1:(C * C)]);
#> }
#> 
#> $model
#> model {
#>   real ps[C, C];
#> 
#>   // Priors
#>   l1_0 ~ normal(0, 2);
#>   l2_0 ~ normal(0, 2);
#>   l3_0 ~ normal(0, 2);
#>   l4_0 ~ normal(0, 2);
#>   l5_0 ~ normal(0, 2);
#>   l6_0 ~ normal(0, 2);
#>   l1_11 ~ lognormal(0, 1);
#>   l2_12 ~ lognormal(0, 1);
#>   l3_11 ~ lognormal(0, 1);
#>   l4_12 ~ lognormal(0, 1);
#>   l5_11 ~ lognormal(0, 1);
#>   l5_12 ~ lognormal(0, 1);
#>   l6_11 ~ lognormal(0, 1);
#>   l6_12 ~ lognormal(0, 1);
#>   l5_212 ~ normal(0, 2);
#>   l6_212 ~ normal(0, 2);
#> 
#>   target += sum(map_rect(sum_probs, beta, theta, xr, xi));
#> }
#> 
#> $generated_quantities
#> generated quantities {
#>   vector[J] log_lik;
#>   matrix[C, C] format_prob_transition_class[J];
#>   vector[J*C*C] prob_transition_class;
#>   matrix[A, 2] prob_resp_attr[J];
#> 
#>   log_lik = map_rect(person_loglik, beta, theta, xr, xi);
#> 
#>   prob_transition_class = map_rect(resp_transition, beta, theta, xr, xi);
#> 
#>   for (j in 1:J) {
#>     for (c1 in 1:C) {
#>       for (c2 in 1:C) {
#>         int iter = ((c1 - 1) * 8) + c2 + ((j - 1) * C * C);
#>         format_prob_transition_class[j,c1,c2] = prob_transition_class[iter];
#>       }
#>     }
#>   }
#> 
#>   for (j in 1:J) {
#>     for (a in 1:A) {
#>       vector[C] prob_attr_class_t1;
#>       vector[C] prob_attr_class_t2;
#>       for (c in 1:C) {
#>         prob_attr_class_t1[c] = sum(format_prob_transition_class[j,c,]) * Alpha[c,a];
#>         prob_attr_class_t2[c] = sum(format_prob_transition_class[j,,c]) * Alpha[c,a];
#>       }
#>       prob_resp_attr[j,a,1] = sum(prob_attr_class_t1);
#>       prob_resp_attr[j,a,2] = sum(prob_attr_class_t2);
#>     }
#>   }
#> }
#>